requirements.txt
openai==0.27.8
load_dotenv==0.1.0